Clarify "more efficient to count down" advice
The advice that "It's almost always more efficient to count
down" is technically correct, but easy to misinterpret.
It's correct in the sense that you almost always encounter
situations where you have a positive number N and you need
to loop N times, so it's more efficient to count down from
N to 0, rather than 0 to N.

But when I first read this passage (and until I carefully
re-read it, and the sample code), I misinterpreted it as
incorrectly saying that counting from -638 to 0 was less
efficient than counting from +638 to 0. And based on that
I misread the sample code as a poorly implemented -638 to 0
loop.

This change aims to clarify the section to avoid
others making the same misinterpretation.