Clarify the note about the off-by-one error
Since the text in this section talks about writing "the same loop", it seemed strange to me when I got to the note saying that the lower loop was written to execute 639 times while the upper loop was written to execute 638 times. I considered changing the lower loop to start at 1 to fix this, but the text also talks about counting from 638-to-0 and 0-to-638, so doing 638-to-0 and 1-to-638 also wouldn't be "the same loop".

I considered rewriting the loop to keep the literals 0 and 638 in the code but only run 638 times, but then I realized that the parts of the loop below the JEZ actually are already running 638 times; it's only the part of the loop above the jump that run 639 times.

So I'm leaving the sample code as it is, and changing the note to clarify that parts of it run 638 times, and parts 639. Also giving the name of the general issue (off-by-one error) because, personally, I had to fix a lot of off-by-ones in my TIS-100 code. :)